Major correction on repositories

See merge request tacos-paza/ms-data!16

+++ src/main/java/com/tacospasa/msdata/Repository/PaymentMethodRepository.java
@@ -5,5 +5,5 @@ import com.tacospasa.msdata.Entity.PaymentMethodEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethodEntity,Integer> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethodEntity,Long> {
 }
